feat: integrar JPA y H2 en las entidades

+++ src/main/java/com/hotel/macondo/controller/HabitacionController.java
@@ -44,7 +44,7 @@ public class HabitacionController {
      * Muestra el detalle de una habitacion o vuelve al listado si no existe.
      */
     @GetMapping("/{id}")
-    public String mostrarHabitacion(@PathVariable("id") Integer id, Model model) {
+    public String mostrarHabitacion(@PathVariable("id") Long id, Model model) {
         Habitacion habitacion = service.buscarPorId(id);
         if (habitacion == null) {
             return "redirect:/habitaciones";
