+++ src/main/java/com/hotel/macondo/service/OperadorService.java
@@ -10,7 +10,7 @@ public interface OperadorService {
     Collection<Operador> buscarTodos();
 
     /** Busca un operador por identificador. */
-    Operador buscarPorId(Integer id);
+    Operador buscarPorId(Long id);
 
     /** Crea o actualiza un operador. */
     Operador guardar(Operador operador);
@@ -22,8 +22,8 @@ public interface OperadorService {
      * Invierte el estado (activo/inactivo) de un operador. Retorna null si el
      * operador no existe.
      */
-    Operador cambiarEstado(Integer id);
+    Operador cambiarEstado(Long id);
 
     /** Elimina un operador por identificador. */
-    void eliminar(Integer id);
+    void eliminar(Long id);
 }
